move mood to its own component
