+++ src/components/mood/mood.scss
